Enable fallback synthetic source for shape
